team_submissions/vivi-sampelan/requirements.txt
langchain==0.1.0
langchain-openai==0.0.5
yfinance==0.2.36
pandas==2.2.0
numpy==1.26.3
python-dotenv==1.0.1
reportlab==4.0.9
matplotlib==3.8.2
